virt-ctrl: Move reactors in testing pkg
