+++ pkg/virt-controller/watch/testing/testing.go
@@ -16,6 +16,7 @@
  * Copyright The KubeVirt Authors
  *
  */
+
 package testing
 
 import (
